Adding basics for pushing and popping columns onto table
The data schema has now been extended to factor in generic channels into record entries. For every page of data that is currently available, all of the available columns that can be displayed can be manually selected to display on the table by ticking its checkbox. This appends it to the table on the right.

The code is a complete mess right now so this will need to be refactored before continuing, otherwise it's just going to be impossible to maintain