DSEGOG-8 prevent reordering when user is resizing
This fixes a bug which caused a column to be dragged whenever it is resized. It was no use trying to disable dragging whenever the column.isResizing field is flagged as true because this only happened after the mouse event triggered, so was always too late. So we use onMouseOver and onMouseOut events instead to enable and disable the behaviour in time.

I spent too long trying to figure out a solution to this