DSEGOG-8 moving checkboxes to table component, fixing column ordering issues
Moving the column checkboxes to be part of the Table component. This makes sense, since we always want the checkboxes to appear alongside the Table and their content should be defined by the Table.

Also redefining column structures and action handlers to better represent user actions and intent. Overall, this improves maintainability and fixes the bug we had whereby columns were incorrectly ordered when a column was removed and added back in.