DSEGOG-10 fix bug with ID header width
ID header can't be closed so we need to not show the close icon. However, this messed with the header width and made it out of sync of the rest of the column. A simple solution to combat this, instead of messing with column widths, was to hide the icon's visibility while still rendering it