DSEGOG-4 fixing data header not sorting properly
Previously, header could only be sorted once and wasn't showing the sort direction. This was due to the dataKey being present in sort object as metadata.dataKey or channels.dataKey.

I've constructed a sortKey in each dataHeader to refer to this property. This may be refactored when Redux is merged in