DSEGOG-4 fixing data race with availableColumns
Sometimes, the collation of the availableColumns field would not occur (presumably because the channels were being constructed from the data gathering request to /records).

I've replaced this with an actual call to /records in fetchChannels(). This means two calls to the /records endpoint but more realistically simulates the website in practice, when we can query an actual channel info endpoint.