DSEGOG-4 reading apiUrl from store, adding preloader for fallback
We now read the apiUrl from the store before sending requests to the backend. However, this requires that all the settings have loaded first so I've added a basic version of the Preloader component from DataGateway that renders while settings are loading. The only difference is that it doesn't spin for some reason. I'll fix this later