Fixing bug with sticky timestamp column
Due to messed-up zIndex values, the column menu and divider was appearing in front of the timestamp column when dragging horizontally. I fixed this by setting the zIndex on the column menu and divider to 0, the same as the parent TableCell. This ensures that column header text remains behind the column menu and the timestamp column is still at the front of the viewport