DSEGOG-52 multiple channel labels in settings
This allows the user to add multiple channels to plot in the plot settings on the y-axis. All this does right now is display a label for each selected channel, this isn't passed to the plot yet. I've also removed the <Select> for choosing a channel, as I feel it gets in the way and the <Autocomplete> does a better job of it anyway