DSEGOG-106 refactoring channels tests
Due to the rework of channels.tsx, I refactored a couple of tests to be more rigorous and imitative of a production scenario.

I've also made a change to generateChannelMetadata to prevent it returning timestamp, shotnum, activeArea and activeExperiment metadata if we have no records in the response. In this scenario, we have no data to present to the user, so it's pointless providing metadata