DSEGOG-69 fixing Enter key bug, improving date passing
Fixing a bug where pressing the Enter key reloaded the page. This was because of the surrounding <form> tag, which I've replaced with a <Grid> to get the same effect.

The picker NEARLY works, but I can't achieve the same functionality with the OK button as in DataGateway. It seems the entirety of the field needs to be completed in the picker before the value is accepted. This is not true if the user inputs the value manually. Still working on this