DSEGOG-69 stopping additional calls when new date matches previous one
With the changes to prevDate, the stopper for onChange being called twice for the same value being passed is gone. This would prevent onChange being called twice on occasions when the same date can be parsed from two values, e.g. 2022-01-01 00:00:0 and 2022-01-01 00:00:00. Therefore, I've added an extra check to prevent this from happening