DSEGOG-202 remove tooltip tests from playwright tests
Due to the more precise way tooltips are now rendered, it is practically impossible to get them to appear in playwright tests anymore. This is because the user needs to precisely line their mouse up over a point to get a tooltip to appear and despite trying to make this happen in tests I couldn't manage it. I was also working blind because the canvas element itself does not appear in playwright trace mode but does in snapshots