DSEGOG-98 DSEGOG-99 change date if both equal, fix copy paste date
I have amended the code to ensure enforcement of code that allows for the date range to be the same date in both from and to fields, thus allowing the user to search down to the exact second.

I have also fixed a bug where pasting a date value into either field didn't work. This was because I wasn't updating the state in the DateTimePicker's onAccept function