DSEGOG-100 add new setting, check incoming count before search query
Adding a new recordLimitWarning setting to operationsgateway-settings.json. When set to an integer, any attempt to fetch more records than this number displays a popup warning message. If this is set to -1, no limit is imposed.

This is done through a new useQuery hook called useIncomingRecordCount. This is called with the potential search parameters and a count is fetched first. If the count is less than the limit or the user presses continue on the popup window, the search is actually performed and the records are fetched. Otherwise, the search is not performed.

Currently, this is only happening for the search bar, will make it happen for filters soon