DSEGOG-100 fixing e2e tests
The search e2e tests need the recordLimitWarning to be set to -1 (hence the stubbed settings response) and an additional wait for an extra count request being made (I'm not sure if this is the incomingRecordCount query or a bug). But for some reason, the filtering e2e tests need none of this to pass, despite using the same functionality. This needs to be investigated