Updating SDK libraries and adding postinstall command
The postinstall command should allow for the editor SDK libraries to be automatically installed whenever a `yarn install` is run. Most users know to run the `yarn install` command before running the app for the first time. Therefore, this means they don't have to worry about needing to know another command to run to be able to start developing the app (if that's what they intend to do)